requirements.txt
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0
rdkit-pypi>=2022.9.3